% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_idf.R
\name{calc_idf}
\alias{calc_idf}
\title{Calculate Inverse Document Frequency (IDF)}
\usage{
calc_idf(tdm)
}
\arguments{
\item{tdm}{A term-document matrix (TDM) where each row represents a
type and each column represents a document.}
}
\value{
A numeric vector containing the Inverse Document Frequency
'DF' for each type in the TDM.
}
\description{
This function is for internal use only.
}
\details{
This function that calculates the Inverse Document Frequency 'IDF'
for each type (e.g., term, lemma) in a term-document matrix (TDM).
It is intended to be used within the package by the
\code{calc_dispersion_metrics} function.
}
\keyword{internal}
