% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confirm_permission.R
\name{confirm_permission}
\alias{confirm_permission}
\title{Confirm permission to use data}
\usage{
confirm_permission()
}
\value{
TRUE if the user confirms permission, FALSE otherwise
}
\description{
This function is for internal use only.
}
\details{
This function confirms that the user has permission to use the data.
If not, the script is returns FALSE and stops.
}
\keyword{internal}
