% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate_enntt_data.R
\name{curate_enntt_data}
\alias{curate_enntt_data}
\title{Curate ENNTT Data}
\usage{
curate_enntt_data(dir_path)
}
\arguments{
\item{dir_path}{A string. The path to the directory containing the data files.}
}
\value{
A data frame. The curated ENNTT data.
}
\description{
This function curates ENNTT data from a specified directory. It searches for files with .dat or .tok extensions,
extracts the unique types from these files, and then curates each type of data using the \code{curate_enntt_file()} function.
The curated data from each type is then combined into a single data frame which is returned.
}
\examples{
\dontrun{
curated_data <- curate_enntt_data('/path/to/data')
}
}
