% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate_enntt_data.R
\name{curate_enntt_file}
\alias{curate_enntt_file}
\title{Curate Single ENNTT File Pair}
\usage{
curate_enntt_file(dir_path, corpus_type)
}
\arguments{
\item{dir_path}{Directory containing the files}

\item{corpus_type}{Type identifier for the corpus}
}
\value{
Data frame of curated data
}
\description{
Curate Single ENNTT File Pair
}
\keyword{internal}
