% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate_swda_data.R
\name{curate_swda_data}
\alias{curate_swda_data}
\title{Curate SWDA data}
\usage{
curate_swda_data(dir_path)
}
\arguments{
\item{dir_path}{The path to the directory containing the .utt files.}
}
\value{
A data frame containing the curated SWDA data.
}
\description{
This function curates the SWDA (Switchboard Dialog Act) data by processing
all the .utt files in the specified directory.
}
\examples{
\dontrun{
curate_swda_data('/path/to/directory')
}
}
