% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate_swda_data.R
\name{curate_swda_file}
\alias{curate_swda_file}
\title{Process a single SWDA utterance file}
\usage{
curate_swda_file(file_path)
}
\arguments{
\item{file_path}{Character string. Path to the .utt file}
}
\value{
A data frame containing processed data from the file
}
\description{
Process a single SWDA utterance file
}
\keyword{internal}
