% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_archive_data.R
\name{download_and_decompress}
\alias{download_and_decompress}
\title{Download and Decompress Archive File}
\usage{
download_and_decompress(url, target_dir, ext)
}
\arguments{
\item{url}{Character string of the archive file URL}

\item{target_dir}{Character string of the target directory path}

\item{ext}{Character string of the file extension}
}
\value{
No return value, called for side effects
}
\description{
Helper function that downloads an archive file to a temporary location
and decompresses it to the target directory.
}
\keyword{internal}
