% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate_enntt_data.R
\name{extract_dat_attrs}
\alias{extract_dat_attrs}
\title{Extract Attributes from XML Line Node}
\usage{
extract_dat_attrs(line_node)
}
\arguments{
\item{line_node}{XML node containing line attributes}
}
\value{
Data frame of extracted attributes
}
\description{
Extract Attributes from XML Line Node
}
\keyword{internal}
