% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate_swda_data.R
\name{extract_speaker_info}
\alias{extract_speaker_info}
\title{Extract speaker information from document lines}
\usage{
extract_speaker_info(doc_lines)
}
\arguments{
\item{doc_lines}{Character vector of file lines}
}
\value{
Named list of speaker information
}
\description{
Extract speaker information from document lines
}
\keyword{internal}
