% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate_swda_data.R
\name{extract_utterances}
\alias{extract_utterances}
\title{Extract and process utterances from document lines}
\usage{
extract_utterances(doc_lines, speaker_info)
}
\arguments{
\item{doc_lines}{Character vector of file lines}

\item{speaker_info}{List of speaker information}
}
\value{
Data frame of processed utterances
}
\description{
Extract and process utterances from document lines
}
\keyword{internal}
