% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate_enntt_data.R
\name{find_enntt_files}
\alias{find_enntt_files}
\title{Find ENNTT Files}
\usage{
find_enntt_files(dir_path)
}
\arguments{
\item{dir_path}{Directory to search for ENNTT files}
}
\value{
Vector of unique corpus types
}
\description{
Find ENNTT Files
}
\keyword{internal}
