% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate_swda_data.R
\name{process_speaker_info}
\alias{process_speaker_info}
\title{Process speaker turn information}
\usage{
process_speaker_info(speaker_turn, speaker_a_id, speaker_b_id)
}
\arguments{
\item{speaker_turn}{Vector of speaker turns}

\item{speaker_a_id}{ID for speaker A}

\item{speaker_b_id}{ID for speaker B}
}
\value{
List with processed speaker information
}
\description{
Process speaker turn information
}
\keyword{internal}
