% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_fig.R
\name{save_fig}
\alias{save_fig}
\title{Save a plot as a file}
\usage{
save_fig(
  plot = NULL,
  file = NULL,
  target_dir = NULL,
  theme = NULL,
  device = "pdf",
  ...
)
}
\arguments{
\item{plot}{The plot to be saved. If not specified, the last plot
created will be saved.}

\item{file}{The name of the file to be saved. If not specified, the
label of the code block will be used.}

\item{target_dir}{The directory where the file will be saved. If not
specified, the current working directory will be used.}

\item{theme}{The ggplot2 theme to be applied to the plot. Default is
the theme specified in the ggplot2 options.}

\item{device}{The device to be used for saving the plot. Options
include "pdf", "png", "jpeg", "tiff", and "svg".}

\item{...}{Additional arguments to be passed to \code{ggsave}.}
}
\value{
The path of the saved file.
}
\description{
This function is a wrapper around \code{ggsave} that allows
you to save a plot as part of a knitr document. It is designed to be
used in a code block. The file name, if not specified, will be the
label of the code block.
}
\examples{
\dontrun{
# Save a plot as a PDF file
p1 <- ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point()
save_fig(p1, file = "mtcars_plot", device = "pdf")
}

}
\keyword{publishing}
