% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_obj.R
\name{save_obj}
\alias{save_obj}
\title{Save an R object as a file}
\usage{
save_obj(object, file = NULL, target_dir = NULL, ...)
}
\arguments{
\item{object}{The R object to be saved.}

\item{file}{The name of the file to be saved. If not specified, the
label of the code block will be used.}

\item{target_dir}{The directory where the file will be saved. If not
specified, the current working directory will be used.}

\item{...}{Additional arguments to be passed to \code{dput}.}
}
\value{
The path of the saved file.
}
\description{
This function is a wrapper around \code{dput} that allows
you to save an R object as part of a knitr document. It is designed to
be used in a code block. The file name, if not specified, will be the
label of the code block.
}
\examples{
\dontrun{
# Save a data frame as a file
save_obj(mtcars, file = "mtcars_data")
}
}
\keyword{publishing}
