% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_tbl.R
\name{save_tbl}
\alias{save_tbl}
\title{Save a table as a file}
\usage{
save_tbl(
  table,
  file = NULL,
  target_dir = NULL,
  bs_theme = "bootstrap",
  device = "pdf",
  ...
)
}
\arguments{
\item{table}{The table to be saved.}

\item{file}{The name of the file to be saved. If not specified,
the name will be based on the current knitr block label.}

\item{target_dir}{The directory where the file will be saved. If not
specified, the current working directory will be used.}

\item{bs_theme}{The Bootstrap theme to be applied to the table (only
applicable for HTML output). Default is "bootstrap".}

\item{device}{The device to be used for saving the table. Options
include "pdf", "html", "latex", "png", and "jpeg". Default is "pdf".}

\item{...}{Additional arguments to be passed to the save_kable function
from the kableExtra package.}
}
\value{
The path of the saved file.
}
\description{
This function is a wrapper around \code{save_kable} from the
kableExtra package that allows you to save a table as part of a knitr
document. It is designed to be used in a code block. The file name,
if not specified, will be the label of the code block.
}
\examples{
\dontrun{
# Save a table as a PDF file
save_tbl(mtcars, file = "mtcars_table", device = "pdf")

# Save a table as an HTML file with a custom Bootstrap theme
save_tbl(mtcars, file = "mtcars_table", device = "html", bs_theme = "flatly")
}

}
\keyword{publishing}
