% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curate_enntt_data.R
\name{validate_dir_path}
\alias{validate_dir_path}
\title{Validate Directory Path}
\usage{
validate_dir_path(dir_path)
}
\arguments{
\item{dir_path}{Directory path to validate}
}
\description{
Validate Directory Path
}
\keyword{internal}
