% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_df.R
\name{calc_df}
\alias{calc_df}
\title{Calculate Document Frequency (DF)}
\usage{
calc_df(tdm)
}
\arguments{
\item{tdm}{A term-document matrix (TDM) where each row represents a type
and each column represents a document.}
}
\value{
A numeric vector containing the Document Frequency 'DF' for each
type in the TDM.
}
\description{
This function is for internal use only.
}
\details{
This function that calculates the Document Frequency 'DF' for each type
(e.g., term, lemma) in a term-document matrix (TDM). It is intended to be
used within the package by the \code{calc_dispersion_metrics} function.
}
\keyword{internal}
