% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_dp.R
\name{calc_dp}
\alias{calc_dp}
\title{Calculate Gries' Deviation of Proportions (DP)}
\usage{
calc_dp(tdm_normalized, corpus_parts)
}
\arguments{
\item{tdm_normalized}{A normalized term-document matrix (TDM) where each
row represents a type and each column represents a document. The values
should be the proportions of each type's frequency to its total frequency
across all documents.}

\item{corpus_parts}{A numeric vector containing the proportions of each
document in the corpus, which is used to calculate the Deviation of
Proportions (DP).}
}
\value{
A numeric vector containing the Deviation of Proportions (DP) for
each type in the TDM.
}
\description{
This function is for internal use only.
}
\details{
This function that calculates the Deviation of Proportions (DP)
based on Gries' Deviation of Proportions method. It is intended to be used
within the package, particularly by the \code{calc_dispersion_metrics} function.
}
\keyword{internal}
