% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_disp.R
\name{calc_rf}
\alias{calc_rf}
\title{Dispersion and frequency functions
These functions are for internal use only.}
\usage{
calc_rf(tdm)
}
\description{
To be used with the \code{calc_type_metrics} function.
Includes dispersion measures: \code{calc_df()} Document Frequency 'DF',
\code{calc_idf()} Inverse Document Frequency 'IDF', and \code{calc_dp()} Gries'
Deviation of Proportions 'DP' functions. Frequency measures include:
\code{calc_rf()} and \code{calc_orf()}. Takes a term-document matrix (TDM) as
input and returns the respective dispersion metrics for each type in the TDM.
}
\keyword{internal}
