% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_origin.R
\name{create_data_origin}
\alias{create_data_origin}
\title{Create data origin file}
\usage{
create_data_origin(file_path, return = FALSE, force = FALSE)
}
\arguments{
\item{file_path}{File path where the data origin file should be saved.}

\item{return}{Logical value indicating whether the data origin should be
returned.}

\item{force}{Logical value indicating whether to overwrite the file if it
already exists.}
}
\value{
A data frame containing the data origin information.
}
\description{
Data frame with attributes about the data origin, written to a CSV file
and optionally returned.
}
\examples{
tmp_file <- tempfile(fileext = ".csv")
create_data_origin(tmp_file)
read.csv(tmp_file)
}
